feat: 관심사 도메인 구독까지 구현 완료 - 추후 원시 타입 / 래퍼 클래스 사이 논의 진행 중

+++ src/main/java/com/lunazkoe/naa/domain/interest/dto/response/InterestDto.java
@@ -8,7 +8,7 @@ public record InterestDto(
         UUID id,
         String name,
         List<String> keywords,
-        Integer subscriberCount,
+        Long subscriberCount,
         Boolean subscribedByMe
 ) {
